# src/main/resources/OpenAPI_ViviGo.yaml
openapi: 3.1.0
info:
  title: ViviGo API
  description: API de ViviGo una aplicación web para la gestión de alojamientos (casas, apartamentos, fincas), reservas y comentarios, con roles diferenciados para usuarios y anfitriones, utilizando Spring Boot (JPA) y Angular.
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
components:
  schemas:
    Usuario:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        contraseña:
          type: string
          minLength: 8
          pattern: '^(?=.*[A-Z])(?=.*[0-9]).{8,}$'
          description: "Mínimo 8 caracteres, debe contener al menos una mayúscula y un número"
        telefono:
          type: string
          minLength: 1
        rol:
          type: string
          enum: [USER, HOST]
        fechaNacimiento:
          type: string
          format: date
        foto:
          type: string
          format: uri
      required:
        - nombre
        - email
        - contraseña
        - telefono
        - rol
        - fechaNacimiento
...
    Alojamiento:
      type: object
      properties:
        titulo:
          type: string
          minLength: 1
        descripcion:
          type: string
          minLength: 1
        ubicacion:
          $ref: '#/components/schemas/Ubicacion'
        precioNoche:
          type: number
          format: float
          minimum: 0
        capacidad:
          type: integer
          minimum: 1
        servicios:
          type: array
          items:
            type: string
        imagenes:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        imagenPrincipal:
          type: integer
          description: "Índice de la imagen principal en el array de imágenes"
          minimum: 0
      required:
        - titulo
        - descripcion
        - ubicacion
        - precioNoche
        - capacidad
        - imagenes
    Ubicacion:
      type: object
      properties:
        ciudad:
          type: string
          minLength: 1
        direccion:
          type: string
          minLength: 1
        coordenadas:
          $ref: '#/components/schemas/Coordenadas'
      required:
        - ciudad
        - direccion
        - coordenadas
    Coordenadas:
      type: object
      properties:
        latitud:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitud:
          type: number
          format: double
          minimum: -180
          maximum: 180
      required:
        - latitud
        - longitud
    Reserva:
      type: object
      properties:
        alojamientoId:
          type: integer
          format: int64
        usuarioId:
          type: integer
          format: int64
        fechaEntrada:
          type: string
          format: date
        fechaSalida:
          type: string
          format: date
        numeroHuespedes:
          type: integer
          minimum: 1
      required:
        - alojamientoId
        - fechaEntrada
        - fechaSalida
        - numeroHuespedes
...
    PaginatedResponse:
      type: object
      properties:
        contenido:
          type: array
          items: {}
        pagina:
          type: integer
          minimum: 0
        tamano:
          type: integer
          minimum: 1
        totalElementos:
          type: integer
          minimum: 0
        totalPaginas:
          type: integer
          minimum: 0
        esUltima:
          type: boolean
        esPrimera:
          type: boolean
...
    Cambio_Contrasena:
      type: object
      properties:
        passwordActual:
          type: string
        passwordNueva:
          type: string
          minLength: 8
          pattern: '^(?=.*[A-Z])(?=.*[0-9]).{8,}$'
      required:
        - passwordActual
        - passwordNueva
    Creacion_Anfitrion:
      type: object
      properties:
        sobreMi:
          type: string
          minLength: 1
        DocumentoLegal:
          type: string
          minLength: 1
      required:
        - sobreMi
        - DocumentoLegal
...
    Error:
      type: object
      properties:
        error:
          type: string
        mensaje:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /usuarios/{id}:
    get:
      summary: Obtener información del usuario
      tags: [ Usuarios ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Datos del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos para obtener el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: El usuario no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Editar perfil del usuario
      tags: [ Usuarios ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                telefono:
                  type: string
                  minLength: 1
                foto:
                  type: string
                  format: uri
                descripcion:
                  type: string
                  description: "Solo para anfitriones"
                documentos:
                  type: array
                  items:
                    type: string
                    format: uri
                  description: "Solo para anfitriones"
      responses:
        '200':
          description: Usuario actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos para actualizar estos datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: El usuario no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Eliminar cuenta del usuario (soft delete)
      tags: [ Usuarios ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Usuario marcado como eliminado
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos para eliminar el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: El usuario no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /usuarios/{id}/contrasena:
    patch:
      summary: Cambiar contraseña (usuario autenticado)
      tags: [ Usuarios ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cambio_Contrasena'
      responses:
        '200':
          description: Contraseña cambiada correctamente
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /usuarios/{id}/alojamientos:
    get:
      summary: Listar alojamientos del anfitrión
      tags: [ Usuarios ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: tamano
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Lista paginada de alojamientos del anfitrión
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      contenido:
                        type: array
                        items:
                          $ref: '#/components/schemas/Alojamiento'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No es anfitrión
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /usuarios/{id}/reservas:
    get:
      summary: Listar reservas del Usuario
      tags: [ Usuarios ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: tamano
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Lista paginada de reservas del usuario
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      contenido:
                        type: array
                        items:
                          $ref: '#/components/schemas/Reserva'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No es el Usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /usuarios/{id}/anfitrion:
    post:
      summary: Convertir usuario a anfitrión
      tags: [ Usuarios ]
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Creacion_Anfitrion'
      responses:
        '200':
          description: Usuario convertido a anfitrión exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: No tiene permisos o usuario no es Huesped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Usuario no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # GESTIÓN DE ALOJAMIENTOS
...